+++ bower.json
@@ -1,28 +1,28 @@
 {
   "name": "ilios",
   "dependencies": {
-    "ember": "1.13.2",
+    "ember": "1.13.5",
     "ember-cli-moment-shim": "~0.0.3",
     "ember-cli-shims": "ember-cli/ember-cli-shims#0.0.3",
     "ember-cli-test-loader": "ember-cli-test-loader#0.1.3",
-    "ember-data": "1.13.3",
+    "ember-data": "1.13.7",
     "ember-inflector": "~1.3.1",
-    "ember-load-initializers": "ember-cli/ember-load-initializers#0.1.4",
-    "ember-qunit": "0.4.0",
+    "ember-load-initializers": "ember-cli/ember-load-initializers#0.1.5",
+    "ember-qunit": "0.4.7",
     "ember-qunit-notifications": "0.0.7",
-    "ember-resolver": "~0.1.15",
+    "ember-resolver": "~0.1.18",
     "jquery": "^1.11.1",
     "loader.js": "ember-cli/loader.js#3.2.0",
     "neat": "~1.7.1",
     "normalize.css": "~3.0.2",
     "pretender": "pretender#master",
-    "qunit": "~1.17.1",
     "pikaday": "1.3.2",
     "moment": "~2.10.3",
     "blanket": "~1.1.5",
     "lodash": "~3.7.0",
     "Faker": "~2.1.3",
     "tooltipster": "~3.3.0",
-    "ember-simple-auth": "0.8.0"
+    "ember-simple-auth": "0.8.0",
+    "qunit": "~1.17.1"
   }
 }
