+++ package.json
@@ -8,12 +8,13 @@
     "test": "tests"
   },
   "scripts": {
+    "start": "ember server",
     "build": "ember build",
     "test": "ember test"
   },
   "repository": "https://github.com/ilios/frontend",
   "engines": {
-    "node": "^0.12.2",
+    "node": "^0.12.5",
     "npm": "2.7.x"
   },
   "author": "",
@@ -23,9 +24,9 @@
     "broccoli-sass": "^0.6.4",
     "codeclimate-test-reporter": "~0.1.0",
     "elemental-calendar": "0.0.7",
-    "ember-cli": "0.2.7",
-    "ember-cli-app-version": "0.3.3",
-    "ember-cli-autoprefixer": "0.4.1",
+    "elemental-calendar": "0.0.6",
+    "ember-cli": "1.13.6",
+    "ember-cli-app-version": "0.4.0",
     "ember-cli-babel": "^5.0.0",
     "ember-cli-blanket": "0.5.3",
     "ember-cli-bourbon": "^1.0.0",
@@ -35,24 +36,26 @@
     "ember-cli-deprecation-workflow": "0.1.3",
     "ember-cli-flash": "1.2.0",
     "ember-cli-font-awesome": "~0.0.9",
-    "ember-cli-htmlbars": "0.7.6",
-    "ember-cli-htmlbars-inline-precompile": "0.1.3",
-    "ember-cli-ic-ajax": "0.1.1",
+    "ember-cli-htmlbars": "0.7.9",
+    "ember-cli-htmlbars-inline-precompile": "^0.1.3",
+    "ember-cli-ic-ajax": "0.2.1",
     "ember-cli-inject-live-reload": "^1.3.0",
     "ember-cli-mirage": "^0.1.4",
     "ember-cli-neat": "0.0.4",
-    "ember-cli-qunit": "0.3.14",
+    "ember-cli-qunit": "0.3.18",
+    "ember-cli-release": "0.2.3",
+    "ember-cli-sri": "^1.0.1",
     "ember-cli-sauce": "^1.0.0",
     "ember-cli-simple-auth": "^0.8.0",
     "ember-cli-simple-auth-testing": "0.8.0",
     "ember-cli-simple-auth-token": "^0.7.3",
     "ember-cli-tooltipster": "0.0.9",
     "ember-cli-uglify": "^1.0.1",
-    "ember-data": "1.13.3",
+    "ember-data": "1.13.7",
     "ember-deploy-s3": "0.0.4",
     "ember-deploy-s3-index": "^0.4.0",
     "ember-disable-proxy-controllers": "^1.0.0",
-    "ember-export-application-global": "^1.0.2",
+    "ember-export-application-global": "^1.0.3",
     "ember-i18n": "^4.1.0",
     "ember-inflector": "^1.6.1",
     "ember-moment": "1.1.1",
