+++ testem.json
@@ -1,6 +1,6 @@
 {
   "framework": "qunit",
-  "test_page": "tests/index.html?coverage",
+  "test_page": "tests/index.html?hidepassed",
   "disable_watching": true,
   "launch_in_ci": [
     "PhantomJS"
